Move WebData and WebDataFilter to mapping package
