+++ src/main/java/de/retest/web/mapping/WebDataFilter.java
@@ -1,4 +1,4 @@
-package de.retest.web;
+package de.retest.web.mapping;
 
 import java.util.Arrays;
 import java.util.HashSet;
